Move directory logic to AppContainer to decouple from googlesheetApi and handle redirect to revere
